Create PyMatch.py
PyMatch is a memory-matching game where you try to find all of the 'cards' with the same symbol on them.  In many versions of this game, there are only two cards of any type but in this version of the game there can be one card, two cards, or more.  If the player picks a card with only one copy, for instance if there were only one card with grapes on them, then that card stays turned over.  If they pick a card that has other copies still to be turned over, then they have to find the rest of those cards and turn them over in order to keep that card turned over.  The game should tell them how many of the cards are left.  If they guess incorrectly, their guesses all turn back over (but previously correct guesses remain turned up).